Add Tools.Common lib, add flat flow test functionality to ReproTool
