+++ arcade-services.sln
@@ -96,6 +96,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ProductConstructionService.
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "FlatFlowMigrationCli", "tools\FlatFlowMigrationCli\FlatFlowMigrationCli.csproj", "{89DF188B-B1FC-D3C4-A76E-019144ABA9CB}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Tools.Common", "tools\Tools.Common\Tools.Common.csproj", "{6D710AEF-A872-4F96-97F1-A8B9845839AB}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -514,6 +516,18 @@ Global
 		{89DF188B-B1FC-D3C4-A76E-019144ABA9CB}.Release|x64.Build.0 = Release|Any CPU
 		{89DF188B-B1FC-D3C4-A76E-019144ABA9CB}.Release|x86.ActiveCfg = Release|Any CPU
 		{89DF188B-B1FC-D3C4-A76E-019144ABA9CB}.Release|x86.Build.0 = Release|Any CPU
+		{6D710AEF-A872-4F96-97F1-A8B9845839AB}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{6D710AEF-A872-4F96-97F1-A8B9845839AB}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{6D710AEF-A872-4F96-97F1-A8B9845839AB}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{6D710AEF-A872-4F96-97F1-A8B9845839AB}.Debug|x64.Build.0 = Debug|Any CPU
+		{6D710AEF-A872-4F96-97F1-A8B9845839AB}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{6D710AEF-A872-4F96-97F1-A8B9845839AB}.Debug|x86.Build.0 = Debug|Any CPU
+		{6D710AEF-A872-4F96-97F1-A8B9845839AB}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{6D710AEF-A872-4F96-97F1-A8B9845839AB}.Release|Any CPU.Build.0 = Release|Any CPU
+		{6D710AEF-A872-4F96-97F1-A8B9845839AB}.Release|x64.ActiveCfg = Release|Any CPU
+		{6D710AEF-A872-4F96-97F1-A8B9845839AB}.Release|x64.Build.0 = Release|Any CPU
+		{6D710AEF-A872-4F96-97F1-A8B9845839AB}.Release|x86.ActiveCfg = Release|Any CPU
+		{6D710AEF-A872-4F96-97F1-A8B9845839AB}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -556,6 +570,7 @@ Global
 		{B95B12A8-FCE1-618A-CA77-134B59A5C050} = {02EA681E-C7D8-13C7-8484-4AC65E1B71E8}
 		{261CB211-6023-8025-48E1-D11953F4C61C} = {02EA681E-C7D8-13C7-8484-4AC65E1B71E8}
 		{89DF188B-B1FC-D3C4-A76E-019144ABA9CB} = {02EA681E-C7D8-13C7-8484-4AC65E1B71E8}
+		{6D710AEF-A872-4F96-97F1-A8B9845839AB} = {02EA681E-C7D8-13C7-8484-4AC65E1B71E8}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {32B9C883-432E-4FC8-A1BF-090EB033DD5B}
