+++ tools/FlatFlowMigrationCli/MigrationLogger.cs
@@ -5,6 +5,7 @@ using System.Text.Json;
 using System.Text.Json.Serialization;
 using Microsoft.DotNet.ProductConstructionService.Client.Models;
 using Microsoft.Extensions.Logging;
+using Tools.Common;
 
 namespace FlatFlowMigrationCli;
