+++ tools/FlatFlowMigrationCli/Operations/MigrateOperation.cs
@@ -9,6 +9,8 @@ using Microsoft.DotNet.DarcLib.VirtualMonoRepo;
 using Microsoft.DotNet.ProductConstructionService.Client;
 using Microsoft.DotNet.ProductConstructionService.Client.Models;
 using Microsoft.Extensions.Logging;
+using Tools.Common;
+using Constants = Tools.Common.Constants;
 
 namespace FlatFlowMigrationCli.Operations;
