+++ tools/FlatFlowMigrationCli/Options/Options.cs
@@ -10,6 +10,7 @@ using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Logging.Console;
 using ProductConstructionService.Common;
+using Tools.Common;
 
 namespace FlatFlowMigrationCli.Options;
