+++ tools/ProductConstructionService.ReproTool/Program.cs
@@ -2,16 +2,26 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 
 using CommandLine;
+using Microsoft.DotNet.DarcLib.VirtualMonoRepo;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using ProductConstructionService.ReproTool;
-using Microsoft.DotNet.DarcLib.VirtualMonoRepo;
+using ProductConstructionService.ReproTool.Operations;
+using ProductConstructionService.ReproTool.Options;
+using Tools.Common;
 
-Parser.Default.ParseArguments<ReproToolOptions>(args)
-    .WithParsed<ReproToolOptions>(o =>
+Type[] options =
+[
+    typeof(ReproOptions),
+    typeof(FlatFlowTestOptions),
+    typeof(FullBackflowTestOptions)
+];
+
+Parser.Default.ParseArguments(args, options)
+    .MapResult((Options o) =>
     {
         IConfiguration userSecrets = new ConfigurationBuilder()
-            .AddUserSecrets<ReproTool>()
+            .AddUserSecrets<ReproOperation>()
             .Build();
         o.GitHubToken ??= userSecrets["GITHUB_TOKEN"];
         o.GitHubToken ??= Environment.GetEnvironmentVariable("GITHUB_TOKEN");
@@ -19,12 +29,15 @@ Parser.Default.ParseArguments<ReproToolOptions>(args)
 
         var services = new ServiceCollection();
 
-        services.RegisterServices(o);
+        o.RegisterServices(services);
         services.AddSingleton<VmrDependencyResolver>();
 
         services.AddMultiVmrSupport(Path.GetTempPath());
 
         var provider = services.BuildServiceProvider();
 
-        ActivatorUtilities.CreateInstance<FlatFlowTestOperation>(provider).TestFlatFlow().GetAwaiter().GetResult();
-    });
+        o.GetOperation(provider).RunAsync().GetAwaiter().GetResult();
+
+        return 0;
+    },
+    (_) => -1);
