+++ tools/Tools.Common/Constants.cs
@@ -1,9 +1,9 @@
 ﻿// Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 
-namespace FlatFlowMigrationCli;
+namespace Tools.Common;
 
-internal static class Constants
+public static class Constants
 {
     public const string VmrUri = "https://github.com/dotnet/dotnet";
     public const string ArcadeRepoUri = "https://github.com/dotnet/arcade";
